@@ -16,23 +16,29 @@
         }
     ],
     "require": {
-        "php": "^5.3.3|^7.0",
+        "php": "^7.1",
         "friendsofsymfony/oauth2-php": "~1.1",
-        "symfony/framework-bundle": "~2.2|~3.0",
-        "symfony/security-bundle": "~2.1|~3.0",
-        "symfony/dependency-injection": "^2.0.5|~3.0"
+        "symfony/dependency-injection": "~3.0|~4.0",
+        "symfony/framework-bundle": "~3.0|~4.0",
+        "symfony/security-bundle": "~3.0|~4.0"
     },
     "require-dev": {
-        "symfony/class-loader": "~2.1|~3.0",
-        "symfony/yaml": "~2.1|~3.0",
-        "symfony/form": "~2.3|~3.0",
-        "symfony/console": "~2.1|~3.0",
-        "willdurand/propel-typehintable-behavior": "^1.0.4",
-        "propel/propel1": "^1.6.5",
-        "phing/phing": "~2.4",
-        "doctrine/mongodb-odm": "~1.0",
         "doctrine/doctrine-bundle": "~1.0",
-        "doctrine/orm": "~2.2"
+        "doctrine/mongodb-odm": "~1.0",
+        "doctrine/orm": "~2.2",
+        "phing/phing": "~2.4",
+        "php-mock/php-mock-phpunit": "~1.0|~2.0",
+        "phpstan/phpstan-phpunit": "~0.9",
+        "phpstan/phpstan-shim": "~0.9",
+        "phpunit/phpunit": "~5.0|~6.0",
+        "propel/propel1": "~1.6",
+        "symfony/class-loader": "~3.0|~4.0",
+        "symfony/console": "~3.0|~4.0",
+        "symfony/form": "~3.0|~4.0",
+        "symfony/phpunit-bridge": "~3.0|~4.0",
+        "symfony/templating": "~3.0|~4.0",
+        "symfony/yaml": "~3.0|~4.0",
+        "willdurand/propel-typehintable-behavior": "~1.0"
     },
     "suggest": {
         "doctrine/doctrine-bundle": "*",
@@ -46,9 +52,12 @@
         "psr-4": { "FOS\\OAuthServerBundle\\": "" },
         "exclude-from-classmap": ["/Tests/"]
     },
+    "config": {
+        "sort-packages": true
+    },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.5-dev"
+            "dev-master": "2.0-dev"
         }
     }
 }
