@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -11,43 +13,219 @@
 
 namespace FOS\OAuthServerBundle\Tests\Entity;
 
-use FOS\OAuthServerBundle\Entity\TokenManager;
+use Doctrine\ORM\AbstractQuery;
+use Doctrine\ORM\EntityManager;
+use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\EntityRepository;
+use Doctrine\ORM\QueryBuilder;
 use FOS\OAuthServerBundle\Entity\AccessToken;
+use FOS\OAuthServerBundle\Entity\TokenManager;
+use FOS\OAuthServerBundle\Model\TokenInterface;
 
-class TokenManagerTest extends \PHPUnit_Framework_TestCase
+/**
+ * @group time-sensitive
+ *
+ * Class TokenManagerTest
+ *
+ * @author Nikola Petkanski <nikola@petkanski.com>
+ */
+class TokenManagerTest extends \PHPUnit\Framework\TestCase
 {
-    protected $em;
+    /**
+     * @var \PHPUnit_Framework_MockObject_MockObject|EntityManagerInterface
+     */
+    protected $entityManager;
 
+    /**
+     * @var \PHPUnit_Framework_MockObject_MockObject|EntityRepository
+     */
     protected $repository;
 
-    protected $class;
+    /**
+     * @var string
+     */
+    protected $className;
 
-    protected $manager;
+    /**
+     * @var TokenManager
+     */
+    protected $instance;
 
     public function setUp()
     {
-        $this->class = 'FOS\OAuthServerBundle\Entity\AccessToken';
-        $this->repository = $this->getMock('Doctrine\ORM\EntityRepository', array(), array(), '', false);
-        $this->em = $this->getMock('Doctrine\ORM\EntityManager', array(), array(), '', false);
-        $this->em->expects($this->once())
-            ->method('getRepository')
-            ->with($this->class)
-            ->will($this->returnValue($this->repository));
+        $this->className = AccessToken::class;
+        $this->repository = $this->getMockBuilder(EntityRepository::class)
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
+        $this->entityManager = $this->getMockBuilder(EntityManager::class)
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
 
-        $this->manager = new TokenManager($this->em, $this->class);
+        $this->entityManager
+            ->expects($this->once())
+            ->method('getRepository')
+            ->with($this->className)
+            ->willReturn($this->repository)
+        ;
+
+        $this->instance = new TokenManager($this->entityManager, $this->className);
+    }
+
+    public function testConstructWillSetParameters()
+    {
+        $this->assertAttributeSame($this->entityManager, 'em', $this->instance);
+        $this->assertAttributeSame($this->repository, 'repository', $this->instance);
+        $this->assertAttributeSame($this->className, 'class', $this->instance);
     }
 
     public function testUpdateTokenPersistsAndFlushes()
     {
         $token = new AccessToken();
 
-        $this->em->expects($this->once())
+        $this->entityManager
+            ->expects($this->once())
             ->method('persist')
-            ->with($token);
-        $this->em->expects($this->once())
-            ->method('flush')
-            ->with();
+            ->with($token)
+        ;
 
-        $this->manager->updateToken($token);
+        $this->entityManager
+            ->expects($this->once())
+            ->method('flush')
+            ->with()
+        ;
+
+        $this->assertNull($this->instance->updateToken($token));
+    }
+
+    public function testGetClass()
+    {
+        $this->assertSame($this->className, $this->instance->getClass());
+    }
+
+    public function testFindTokenBy()
+    {
+        $randomResult = \random_bytes(5);
+
+        $criteria = [
+            \random_bytes(5),
+        ];
+
+        $this->repository
+            ->expects($this->once())
+            ->method('findOneBy')
+            ->with($criteria)
+            ->willReturn($randomResult)
+        ;
+
+        $this->assertSame($randomResult, $this->instance->findTokenBy($criteria));
+    }
+
+    public function testUpdateToken()
+    {
+        $token = $this->getMockBuilder(TokenInterface::class)
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
+
+        $this->entityManager
+            ->expects($this->once())
+            ->method('persist')
+            ->with($token)
+            ->willReturn(null)
+        ;
+
+        $this->entityManager
+            ->expects($this->once())
+            ->method('flush')
+            ->with()
+            ->willReturn(null)
+        ;
+
+        $this->assertNull($this->instance->updateToken($token));
+    }
+
+    public function testDeleteToken()
+    {
+        $token = $this->getMockBuilder(TokenInterface::class)
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
+
+        $this->entityManager
+            ->expects($this->once())
+            ->method('remove')
+            ->with($token)
+            ->willReturn(null)
+        ;
+
+        $this->entityManager
+            ->expects($this->once())
+            ->method('flush')
+            ->with()
+            ->willReturn(null)
+        ;
+
+        $this->assertNull($this->instance->deleteToken($token));
+    }
+
+    public function testDeleteExpired()
+    {
+        $randomResult = \random_bytes(10);
+
+        $queryBuilder = $this->getMockBuilder(QueryBuilder::class)
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
+
+        $this->repository
+            ->expects($this->once())
+            ->method('createQueryBuilder')
+            ->with('t')
+            ->willReturn($queryBuilder)
+        ;
+
+        $queryBuilder
+            ->expects($this->once())
+            ->method('delete')
+            ->with()
+            ->willReturn($queryBuilder)
+        ;
+
+        $queryBuilder
+            ->expects($this->once())
+            ->method('where')
+            ->with('t.expiresAt < ?1')
+            ->willReturn($queryBuilder)
+        ;
+
+        $queryBuilder
+            ->expects($this->once())
+            ->method('setParameters')
+            ->with([1 => time()])
+            ->willReturn($queryBuilder)
+        ;
+
+        $query = $this->getMockBuilder(AbstractQuery::class)
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
+
+        $queryBuilder
+            ->expects($this->once())
+            ->method('getQuery')
+            ->with()
+            ->willReturn($query)
+        ;
+
+        $query
+            ->expects($this->once())
+            ->method('execute')
+            ->with()
+            ->willReturn($randomResult)
+        ;
+
+        $this->assertSame($randomResult, $this->instance->deleteExpired());
     }
 }
