@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -11,24 +13,61 @@
 
 namespace FOS\OAuthServerBundle\Tests\Security\Authentication\Provider;
 
+use FOS\OAuthServerBundle\Model\AccessToken;
 use FOS\OAuthServerBundle\Security\Authentication\Provider\OAuthProvider;
 use FOS\OAuthServerBundle\Security\Authentication\Token\OAuthToken;
-use FOS\OAuthServerBundle\Model\AccessToken;
+use OAuth2\OAuth2;
+use Symfony\Component\Security\Core\Role\Role;
+use Symfony\Component\Security\Core\User\UserCheckerInterface;
+use Symfony\Component\Security\Core\User\UserInterface;
+use Symfony\Component\Security\Core\User\UserProviderInterface;
 
-class OAuthProviderTest extends \PHPUnit_Framework_TestCase
+class OAuthProviderTest extends \PHPUnit\Framework\TestCase
 {
+    /**
+     * @var \PHPUnit_Framework_MockObject_MockObject|UserInterface
+     */
     protected $user;
+
+    /**
+     * @var \PHPUnit_Framework_MockObject_MockObject|UserProviderInterface
+     */
     protected $userProvider;
+
+    /**
+     * @var OAuthProvider
+     */
     protected $provider;
+
+    /**
+     * @var \PHPUnit_Framework_MockObject_MockObject|OAuth2
+     */
     protected $serverService;
+
+    /**
+     * @var UserCheckerInterface
+     */
     protected $userChecker;
 
     public function setUp()
     {
-        $this->user = $this->getMock('Symfony\Component\Security\Core\User\UserInterface');
-        $this->userProvider = $this->getMock('Symfony\Component\Security\Core\User\UserProviderInterface');
-        $this->serverService = $this->getMock('OAuth2\OAuth2', array('verifyAccessToken'), array(), '', false);
-        $this->userChecker = $this->getMock('Symfony\Component\Security\Core\User\UserCheckerInterface');
+        $this->user = $this->getMockBuilder(UserInterface::class)
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
+        $this->userProvider = $this->getMockBuilder(UserProviderInterface::class)
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
+        $this->serverService = $this->getMockBuilder(OAuth2::class)
+            ->disableOriginalConstructor()
+            ->setMethods(['verifyAccessToken'])
+            ->getMock()
+        ;
+        $this->userChecker = $this->getMockBuilder(UserCheckerInterface::class)
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
         $this->provider = new OAuthProvider($this->userProvider, $this->serverService, $this->userChecker);
     }
 
@@ -39,7 +78,8 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
 
         $this->user->expects($this->once())
             ->method('getRoles')
-            ->will($this->returnValue(array('ROLE_USER')));
+            ->will($this->returnValue(['ROLE_USER']))
+        ;
 
         $accessToken = new AccessToken();
         $accessToken->setUser($this->user);
@@ -47,17 +87,18 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $this->serverService->expects($this->once())
             ->method('verifyAccessToken')
             ->with('x')
-            ->will($this->returnValue($accessToken));
+            ->will($this->returnValue($accessToken))
+        ;
 
         $result = $this->provider->authenticate($token);
 
         $this->assertSame($this->user, $result->getUser());
-        $this->assertEquals($token->getToken(), $result->getToken());
+        $this->assertSame($token->getToken(), $result->getToken());
         $this->assertTrue($result->isAuthenticated());
         $this->assertCount(1, $result->getRoles());
 
         $roles = $result->getRoles();
-        $this->assertEquals('ROLE_USER', $roles[0]->getRole());
+        $this->assertSame('ROLE_USER', $roles[0]->getRole());
     }
 
     public function testAuthenticateReturnsTokenIfValidEvenIfNullData()
@@ -70,7 +111,8 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $this->serverService->expects($this->once())
             ->method('verifyAccessToken')
             ->with('x')
-            ->will($this->returnValue($accessToken));
+            ->will($this->returnValue($accessToken))
+        ;
 
         $result = $this->provider->authenticate($token);
 
@@ -90,7 +132,8 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $this->serverService->expects($this->once())
             ->method('verifyAccessToken')
             ->with('x')
-            ->will($this->returnValue($accessToken));
+            ->will($this->returnValue($accessToken))
+        ;
 
         $result = $this->provider->authenticate($token);
 
@@ -99,10 +142,10 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
 
         $roles = $result->getRoles();
         $this->assertCount(2, $roles);
-        $this->assertInstanceOf('Symfony\Component\Security\Core\Role\Role', $roles[0]);
-        $this->assertEquals('ROLE_FOO', $roles[0]->getRole());
-        $this->assertInstanceOf('Symfony\Component\Security\Core\Role\Role', $roles[1]);
-        $this->assertEquals('ROLE_BAR', $roles[1]->getRole());
+        $this->assertInstanceOf(Role::class, $roles[0]);
+        $this->assertSame('ROLE_FOO', $roles[0]->getRole());
+        $this->assertInstanceOf(Role::class, $roles[1]);
+        $this->assertSame('ROLE_BAR', $roles[1]->getRole());
     }
 
     public function testAuthenticateWithNullScope()
@@ -116,7 +159,8 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $this->serverService->expects($this->once())
             ->method('verifyAccessToken')
             ->with('x')
-            ->will($this->returnValue($accessToken));
+            ->will($this->returnValue($accessToken))
+        ;
 
         $result = $this->provider->authenticate($token);
 
@@ -138,7 +182,8 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $this->serverService->expects($this->once())
             ->method('verifyAccessToken')
             ->with('x')
-            ->will($this->returnValue($accessToken));
+            ->will($this->returnValue($accessToken))
+        ;
 
         $result = $this->provider->authenticate($token);
 
