@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -13,8 +15,8 @@ namespace FOS\OAuthServerBundle\Document;
 
 use Doctrine\ODM\MongoDB\DocumentManager;
 use Doctrine\ODM\MongoDB\DocumentRepository;
-use FOS\OAuthServerBundle\Model\ClientManager as BaseClientManager;
 use FOS\OAuthServerBundle\Model\ClientInterface;
+use FOS\OAuthServerBundle\Model\ClientManager as BaseClientManager;
 
 class ClientManager extends BaseClientManager
 {
@@ -35,8 +37,12 @@ class ClientManager extends BaseClientManager
 
     public function __construct(DocumentManager $dm, $class)
     {
+        // NOTE: bug in Doctrine, hinting DocumentRepository|ObjectRepository when only DocumentRepository is expected
+        /** @var DocumentRepository $repository */
+        $repository = $dm->getRepository($class);
+
         $this->dm = $dm;
-        $this->repository = $dm->getRepository($class);
+        $this->repository = $repository;
         $this->class = $class;
     }
 
