@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -13,15 +15,14 @@ namespace FOS\OAuthServerBundle\Tests\Functional;
 
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\HttpKernel\KernelInterface;
 
 abstract class TestCase extends WebTestCase
 {
-    protected static function createKernel(array $options = array())
-    {
-        $env = @$options['env'] ?: 'test';
-
-        return new AppKernel($env, true);
-    }
+    /**
+     * @var null|KernelInterface
+     */
+    protected static $kernel;
 
     protected function setUp()
     {
@@ -32,5 +33,12 @@ abstract class TestCase extends WebTestCase
     protected function tearDown()
     {
         static::$kernel = null;
+    }
+
+    protected static function createKernel(array $options = [])
+    {
+        $env = @$options['env'] ?: 'test';
+
+        return new AppKernel($env, true);
     }
 }
