@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -15,6 +17,11 @@ use FOS\OAuthServerBundle\Propel\RefreshToken as Token;
 use FOS\OAuthServerBundle\Propel\RefreshTokenQuery as TokenQuery;
 use FOS\OAuthServerBundle\Propel\TokenManager;
 
+/**
+ * @group time-sensitive
+ *
+ * Class TokenManagerTest
+ */
 class TokenManagerTest extends PropelTestCase
 {
     const TOKEN_CLASS = 'FOS\OAuthServerBundle\Propel\RefreshToken';
@@ -31,7 +38,7 @@ class TokenManagerTest extends PropelTestCase
 
     public function testConstruct()
     {
-        $this->assertEquals(self::TOKEN_CLASS, $this->manager->getClass());
+        $this->assertSame(self::TOKEN_CLASS, $this->manager->getClass());
     }
 
     public function testCreateClass()
@@ -42,47 +49,55 @@ class TokenManagerTest extends PropelTestCase
 
     public function testUpdate()
     {
-        $token = $this->getMock('FOS\OAuthServerBundle\Propel\Token');
+        $token = $this->getMockBuilder('FOS\OAuthServerBundle\Propel\Token')
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
         $token
             ->expects($this->once())
-            ->method('save');
+            ->method('save')
+        ;
 
         $this->manager->updateToken($token);
     }
 
     public function testDelete()
     {
-        $token = $this->getMock('FOS\OAuthServerBundle\Propel\Token');
+        $token = $this->getMockBuilder('FOS\OAuthServerBundle\Propel\Token')
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
         $token
             ->expects($this->once())
-            ->method('delete');
+            ->method('delete')
+        ;
 
         $this->manager->deleteToken($token);
     }
 
     public function testFindTokenReturnsNullIfNotFound()
     {
-        $token = $this->manager->findTokenBy(array('token' => '12345'));
+        $token = $this->manager->findTokenBy(['token' => '12345']);
 
         $this->assertNull($token);
     }
 
     public function testFindTokenWithInvalidData()
     {
-        $token = $this->manager->findTokenBy(array('foo' => '12345'));
+        $token = $this->manager->findTokenBy(['foo' => '12345']);
         $this->assertNull($token);
 
-        $token = $this->manager->findTokenBy(array());
+        $token = $this->manager->findTokenBy([]);
         $this->assertNull($token);
 
-        $token = $this->manager->findTokenBy(array('token'));
+        $token = $this->manager->findTokenBy(['token']);
         $this->assertNull($token);
     }
 
     public function testFindToken()
     {
         $token = $this->createToken('12345');
-        $return = $this->manager->findTokenBy(array('token' => '12345'));
+        $return = $this->manager->findTokenBy(['token' => '12345']);
 
         $this->assertNotNull($return);
         $this->assertSame($token, $return);
@@ -109,11 +124,11 @@ class TokenManagerTest extends PropelTestCase
         $a1 = $this->createToken('12345', time() + 100);
         $a2 = $this->createToken('67890', time() - 100);
 
-        $this->assertEquals(2, TokenQuery::create()->count());
+        $this->assertSame(2, TokenQuery::create()->count());
 
         $nb = $this->manager->deleteExpired();
 
-        $this->assertEquals(1, $nb);
+        $this->assertSame(1, $nb);
         $this->assertTrue($a1->equals(TokenQuery::create()->findOne()));
     }
 
