@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -11,8 +13,13 @@
 
 namespace FOS\OAuthServerBundle\Tests\Security\Firewall;
 
+use FOS\OAuthServerBundle\Security\Authentication\Token\OAuthToken;
 use FOS\OAuthServerBundle\Security\Firewall\OAuthListener;
 use FOS\OAuthServerBundle\Tests\TestCase;
+use OAuth2\OAuth2;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Event\GetResponseEvent;
+use Symfony\Component\Security\Core\Authentication\AuthenticationManagerInterface;
 
 class OAuthListenerTest extends TestCase
 {
@@ -26,26 +33,35 @@ class OAuthListenerTest extends TestCase
 
     public function setUp()
     {
-        $this->serverService = $this
-            ->getMockBuilder('OAuth2\OAuth2')
+        $this->serverService = $this->getMockBuilder(OAuth2::class)
             ->disableOriginalConstructor()
-            ->getMock();
+            ->getMock()
+        ;
 
         $this->authManager = $this
-            ->getMock('Symfony\Component\Security\Core\Authentication\AuthenticationManagerInterface');
+            ->getMockBuilder(AuthenticationManagerInterface::class)
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
 
         if (interface_exists('Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface')) {
             $this->securityContext = $this
-                ->getMock('Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface');
+                ->getMockBuilder('Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface')
+                ->disableOriginalConstructor()
+                ->getMock()
+            ;
         } else {
-            $this->securityContext = $this
-                ->getMock('Symfony\Component\Security\Core\SecurityContextInterface');
+            $this->securityContext = $this->getMockBuilder('Symfony\Component\Security\Core\SecurityContextInterface')
+                ->disableOriginalConstructor()
+                ->getMock()
+            ;
         }
 
         $this->event = $this
-            ->getMockBuilder('Symfony\Component\HttpKernel\Event\GetResponseEvent')
+            ->getMockBuilder(GetResponseEvent::class)
             ->disableOriginalConstructor()
-            ->getMock();
+            ->getMock()
+        ;
     }
 
     public function testHandle()
@@ -55,22 +71,26 @@ class OAuthListenerTest extends TestCase
         $this->serverService
             ->expects($this->once())
             ->method('getBearerToken')
-            ->will($this->returnValue('a-token'));
+            ->will($this->returnValue('a-token'))
+        ;
 
         $this->authManager
             ->expects($this->once())
             ->method('authenticate')
-            ->will($this->returnArgument(0));
+            ->will($this->returnArgument(0))
+        ;
 
         $this->securityContext
             ->expects($this->once())
             ->method('setToken')
-            ->will($this->returnArgument(0));
+            ->will($this->returnArgument(0))
+        ;
 
+        /** @var OAuthToken $token */
         $token = $listener->handle($this->event);
 
-        $this->assertInstanceOf('FOS\OAuthServerBundle\Security\Authentication\Token\OAuthToken', $token);
-        $this->assertEquals('a-token', $token->getToken());
+        $this->assertInstanceOf(OAuthToken::class, $token);
+        $this->assertSame('a-token', $token->getToken());
     }
 
     public function testHandleResponse()
@@ -80,26 +100,33 @@ class OAuthListenerTest extends TestCase
         $this->serverService
             ->expects($this->once())
             ->method('getBearerToken')
-            ->will($this->returnValue('a-token'));
+            ->will($this->returnValue('a-token'))
+        ;
 
-        $response = $this->getMock('Symfony\Component\HttpFoundation\Response');
+        $response = $this->getMockBuilder(Response::class)
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
 
         $this->authManager
             ->expects($this->once())
             ->method('authenticate')
-            ->will($this->returnValue($response));
+            ->will($this->returnValue($response))
+        ;
 
         $this->securityContext
             ->expects($this->never())
-            ->method('setToken');
+            ->method('setToken')
+        ;
 
         $this->event
             ->expects($this->once())
             ->method('setResponse')
-            ->will($this->returnArgument(0));
+            ->will($this->returnArgument(0))
+        ;
 
         $ret = $listener->handle($this->event);
 
-        $this->assertEquals($response, $ret);
+        $this->assertSame($response, $ret);
     }
 }
